More changes for IE compatibility
Addresses BitBucket issue #1

Further testing in IE 8 showed that more elements needed to be created manually instead of by jQuery/jqLite.

Also updated the README with steps for IE 8 support.  IE 6 and 7 have limited support, even for Angular
itself.  Neither version will be targeted in the near future.  IE 8 support is still not a guarantee,
especially since Angular will only partially support it themselves, but it is much stronger than it was.

Updated the demo project with the IE 8 "hacks" needed to get it working in that browser.

Signed-off-by: Daniel Hunsaker <danhunsaker@gmail.com>