Changed template to accept arrays of objects
This allows you to add multiple sets of fields to a single template property.  This comes in handy if you want to have multiple check-lists in the same property to have sub sections.  This probably won't work well for other types of fields like radios and such as the values are still stored under the single property name.